# src/main/resources/static/docs/openapi.yaml
openapi: "3.0.0"
info:
  title: ZOO System
  description: API para gestión de zoologicos. Desarrollada con Spring boot y PostgreSQL.
  contact:
    name: Brayan Barco
    url: alex.barco.maicol@gmail.com
    email: alex.barco.maicol@gmail.com
  version: 1.0.0
  x-products:
    - id: zoo-api
      name: ZOO System
      description: API de gestión de Zoologicos
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
  - url: http://localhost:8080
    description: Servidor de desarrollo
  - url: http://localhost:8080
    description: Servidor de QA
  - url: http://localhost:8080
    description: Servidor de producción
security:
  - Security_Token: []
paths:
...
  /api/animals/{id}:
    get:
      tags:
        - animal-controller
      summary: Obtener animal
      description: Obtener detalles de un animal específico
      operationId: getAnimalById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOAnimalResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    put:
      tags:
        - animal-controller
      summary: Actualizar animal
      description: Actualizar información de animal (solo Admin)
      operationId: updateAnimal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimalRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOAnimalResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    delete:
      tags:
        - animal-controller
      summary: Eliminar animal
      description: Eliminar animal y sus comentarios (solo Admin)
      operationId: deleteAnimal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOVoid'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
...
  /api/comments:
    post:
      tags:
        - comment-controller
      summary: Crear comentario
      description: Agregar comentario a un animal (Admin y Empleados)
      operationId: addCommentToAnimal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
        required: true
      responses:
        '201':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /api/comments/{commentId}/replies:
    post:
      tags:
        - comment-controller
      summary: Responder comentario
      description: Agregar respuesta a un comentario existente
      operationId: addReplyToComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /api/animals:
    get:
      tags:
        - animal-controller
      summary: Listar animales
      description: Obtener todos los animales registrados
      operationId: getAllAnimals
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOListAnimalResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    post:
      tags:
        - animal-controller
      summary: Crear animal
      description: Registrar nuevo animal (solo Admin)
      operationId: createAnimal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimalRequest'
        required: true
      responses:
        '201':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOAnimalResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
...
  /api/stats/animals-by-date:
    get:
      tags:
        - stats-controller
      summary: Animales por fecha
      description: Listar animales registrados en fecha específica con detalles
      operationId: getAnimalsByDate
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnimalDetailResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /api/comments/{commentId}:
    get:
      tags:
        - comment-controller
      summary: Obtener comentario
      description: Obtener comentario con sus respuestas
      operationId: getCommentWithReplies
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
    delete:
      tags:
        - comment-controller
      summary: Eliminar comentario
      description: Eliminar comentario y sus respuestas (Admin puede borrar cualquier
      operationId: deleteComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /api/comments/animal/{animalId}:
    get:
      tags:
        - comment-controller
      summary: Comentarios de animal
      description: Obtener todos los comentarios de un animal específico
      operationId: getCommentsForAnimal
      parameters:
        - name: animalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
  /api/animals/by-date:
    get:
      tags:
        - animal-controller
      summary: Animales por fecha
      description: Listar animales registrados en una fecha específica
      operationId: getAnimalsByDate_1
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SuccessResponseDTOListAnimalResponse'
        '400':
          $ref: '#/components/responses/badRequestResponse'
        '401':
          $ref: '#/components/responses/unauthorizedResponse'
        '403':
          $ref: '#/components/responses/forbiddenResponse'
        '404':
          $ref: '#/components/responses/notFoundResponse'
        '406':
          $ref: '#/components/responses/notAcceptableResponse'
        '413':
          $ref: '#/components/responses/payloadTooLargeResponse'
        '415':
          $ref: '#/components/responses/unsupportedMediaTypeResponse'
        '429':
          $ref: '#/components/responses/tooManyRequestsResponse'
        '500':
          $ref: '#/components/responses/internalServerErrorResponse'
        '503':
          $ref: '#/components/responses/serviceUnavailableResponse'
components:
  schemas:
    StandardError:
      type: object
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 20
          description: Código del error
          example: PERR100
        description:
          type: string
          minLength: 2
          maxLength: 250
          description: Descripción del error
          example: Error de validación
        field:
          type: string
          minLength: 2
          maxLength: 100
          description: Campo con error
          example: email
        traceId:
          type: string
          format: uuid
          description: ID de traza para diagnóstico
          example: "550e8400-e29b-41d4-a716-446655440000"
      required:
        - code
        - description
    StandardErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/StandardError'
...
    AnimalRequest:
      type: object
      properties:
        name:
          type: string
        speciesId:
          type: integer
          format: int64
        zoneId:
          type: integer
          format: int64
    AnimalResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        species:
          $ref: '#/components/schemas/SpeciesSimpleResponse'
        zone:
          $ref: '#/components/schemas/ZoneSimpleResponse'
        registrationDate:
          type: string
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentSimpleResponse'
    CommentSimpleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
        author:
          type: string
        createdAt:
          type: string
    SpeciesSimpleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SuccessResponseDTOAnimalResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AnimalResponse'
        errors: {}
    ZoneSimpleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    CommentRequest:
      type: object
      properties:
        animalId:
          type: integer
          format: int64
        message:
          type: string
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
        author:
          type: string
        createdAt:
          type: string
        animalId:
          type: integer
          format: int64
        parentCommentId:
          type: integer
          format: int64
        replies:
          type: array
          items: {}
...
    AnimalDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        species:
          type: string
        zone:
          type: string
        registrationDate:
          type: string
...
    SuccessResponseDTOListAnimalResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AnimalResponse'
        errors: {}
    SuccessResponseDTOVoid:
      type: object
      properties:
        data: {}
        errors: {}
  securitySchemes:
    Security_Token:
      type: http
      description: Access Token For My API
      scheme: bearer
      bearerFormat: JWT
  responses:
    badRequestResponse:
      description: Solicitud incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            BadRequest:
              value:
                errors:
                  - code: "400"
                    description: "El campo ' ' es requerido"
                    field: "email"
                    traceId: "550e8400-e29b-41d4-a716-446655440000"
    unauthorizedResponse:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            Unauthorized:
              value:
                errors:
                  - code: "401"
                    description: "Token de acceso no válido o expirado"
    forbiddenResponse:
      description: Prohibido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            Forbidden:
              value:
                errors:
                  - code: "403"
                    description: "No tiene permisos para acceder a este recurso"
    notFoundResponse:
      description: No encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            NotFound:
              value:
                errors:
                  - code: "404"
                    description: "Elemento no encontrado"
    internalServerErrorResponse:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            InternalError:
              value:
                errors:
                  - code: "500"
                    description: "Error interno del servidor"
    notAcceptableResponse:
      description: No aceptable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            NotAcceptable:
              value:
                errors:
                  - code: "406"
                    description: "El encabezado Accept no incluye un tipo de medio soportado"
    payloadTooLargeResponse:
      description: Carga útil demasiado grande
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            PayloadTooLarge:
              value:
                errors:
                  - code: "413"
                    description: "La carga útil de la solicitud excede el límite permitido"
    unsupportedMediaTypeResponse:
      description: Tipo de medio no soportado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            UnsupportedMediaType:
              value:
                errors:
                  - code: "415"
                    description: "El tipo de medio de la solicitud no está soportado"
    tooManyRequestsResponse:
      description: Demasiadas solicitudes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            TooManyRequests:
              value:
                errors:
                  - code: "429"
                    description: "Se ha excedido el límite de tasa de solicitudes"
    serviceUnavailableResponse:
      description: Servicio no disponible
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          examples:
            ServiceUnavailable:
              value:
                errors:
                  - code: "503"
                    description: "El servicio no está disponible temporalmente"